Move files back to Central.
