+++ central/apitoken/service/privilege_escalation_test.go
@@ -3,9 +3,9 @@ package service
 import (
 	"testing"
 
+	"github.com/stackrox/rox/central/role"
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/auth/permissions"
-	"github.com/stackrox/rox/pkg/auth/role"
 	"github.com/stackrox/rox/pkg/testutils/roletest"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
