+++ central/sac/authorizer/builtin_scoped_authorizer_test.go
@@ -5,10 +5,10 @@ import (
 	"testing"
 
 	"github.com/stackrox/default-authz-plugin/pkg/payload"
+	rolePkg "github.com/stackrox/rox/central/role"
 	"github.com/stackrox/rox/central/role/resources"
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/auth/permissions"
-	rolePkg "github.com/stackrox/rox/pkg/auth/role"
 	"github.com/stackrox/rox/pkg/buildinfo"
 	"github.com/stackrox/rox/pkg/sac"
 	"github.com/stackrox/rox/pkg/sac/observe"
